Restructure project and split off SearchFlightClient interface.
