+++ src/main/java/eu/isabel/interview/App.java
@@ -1,11 +1,11 @@
 package eu.isabel.interview;
 
-import static eu.isabel.interview.task1.Airport.londonGatwick;
-import static eu.isabel.interview.task1.Airport.parisOrly;
+import static eu.isabel.interview.domain.models.Airport.londonGatwick;
+import static eu.isabel.interview.domain.models.Airport.parisOrly;
 
-import eu.isabel.interview.task1.Flight;
-import eu.isabel.interview.task1.SearchFlightClient.SearchFlightClientMock;
-import eu.isabel.interview.task1.SearchFlightService.SearchFlightServiceFactory;
+import eu.isabel.interview.domain.models.Flight;
+import eu.isabel.interview.infrastructure.clients.SearchFlightClientMock;
+import eu.isabel.interview.application.services.SearchFlightService.SearchFlightServiceFactory;
 import java.time.Duration;
 import java.time.Instant;
 import java.time.LocalDate;
